requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.60.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
email-validator>=2.0.0
psycopg2-binary>=2.9.0
